feat: Добавить MarkDbStorage, MarkDao. Добавить методы saveMark, updateMark, deleteMark в FilmService. Добавить таблицу film_marks.

+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -35,5 +35,5 @@ public class Film {
     @NotNull
     private Mpa mpa;
     @JsonIgnore
-    private Integer rate = 0;
+    private Float rate = 0f;
 }
